first upload
my first vb program that saves its settings properly on close. better late than never :P

this program is made in a very old version of visual studio for nostalgia' sake, sorry in advance if that results in an annoying file structure.

a known issue with this program in its current state is not behaving properly if you force the application to manage high DPI through Windows. this is not a problem for me since i always use 100% but i do want to revisit it in a future version.